// packages/webdriver-image-comparison/src/helpers/options.interface.ts
export enum LogLevel {
    debug = 'debug',
    info = 'info',
    warn = 'warn',
    silent = 'silent',
}